Xvideos video parser.

+++ shared/src/commonMain/kotlin/com/yaikostudio/kparsetron/entities/users/Account.kt
@@ -8,5 +8,5 @@ data class Account(
     val username: String,
     val isVerified: Boolean,
     val profilePicUrl: ResourceUrl?,
-    val followers: Int?,
+    val followers: Long?,
 )
